requirements.txt
asgiref==3.8.1
Django==5.0.4
django-localflavor==4.0
psycopg2==2.9.9
python-decouple==3.8
python-stdnum==1.20
sqlparse==0.5.0
tzdata==2024.1
